+++ sources/packages/web/src/constants/routes/RouteConstants.ts
@@ -79,7 +79,7 @@ export const InstitutionRoutesConst = {
   STUDENT_APPLICATION_DETAILS: Symbol(),
   STUDENT_RESTRICTIONS: Symbol(),
   STUDENT_FILE_UPLOADS: Symbol(),
-  STUDENT_OVERAWARDS: Symbol(),
+  STUDENT_BALANCES: Symbol(),
   STUDENT_NOTES: Symbol(),
   ASSESSMENTS_SUMMARY: Symbol(),
   APPLICATION_EXCEPTION: Symbol(),
@@ -108,7 +108,7 @@ export const AESTRoutesConst = {
   STUDENT_FILE_UPLOADS: Symbol(),
   STUDENT_NOTES: Symbol(),
   SIN_MANAGEMENT: Symbol(),
-  STUDENT_OVERAWARDS: Symbol(),
+  STUDENT_BALANCES: Symbol(),
   PROGRAM_DETAILS: Symbol(),
   SEARCH_INSTITUTIONS: Symbol(),
   INSTITUTION_PROFILE: Symbol(),
